Use chocolatey for install software in Vagrant boxes
Using msi directly to install software and make sure we have the correct
dependencies is very brittle and increases maintenance cost, especially
when compared to Linux. Use [chocolatey](https://chocolatey.org/)
package manage to install all the software we need for development.

This makes our provisioning script much simpler and a lot more robust
since we leverage the work of the Windows community on what is the best
way to install the software.

Provisioning logs:
- [Windows Server](https://gitlab.com/snippets/1942359)
- [Windows 10](https://gitlab.com/snippets/1942363)

Closes https://gitlab.com/gitlab-org/gitlab-runner/issues/8309