Fix millicpu comparison for maxOverwrite
Bug:
Inside of the config.toml the user set the values of
[`cpu_limit_overwrite_max_allowed`](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/a819b26dfd0a23dfbfbc57925233f2e3b6f87635/common/config.go#L172)
and
[`cpu_request_overwrite_max_allowed`](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/a819b26dfd0a23dfbfbc57925233f2e3b6f87635/common/config.go#L180)
using the
[millicpu](https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/#meaning-of-cpu)
such as `400m`. However, if the user set the `KUBERNETES_CPU_REQUEST` to
something higher such as `600m` the request is still applied when it
shouldn't because `600m` is higher than `400m`. This is because of the
overwrite is [transformed using
.Value](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/a819b26dfd0a23dfbfbc57925233f2e3b6f87635/executors/kubernetes/overwrites.go#L202)
which rounds up so both `600m` and `400m` are rounded up to `1` which
end up being the same which is not correct. Use
[cmp](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/a819b26dfd0a23dfbfbc57925233f2e3b6f87635/vendor/k8s.io/apimachinery/pkg/api/resource/quantity.go#L559-566)
instead, that is resilient to such changes.

Extra:
- Update the error message for when the resource is higher to be more
  accurate `is higher than the limit allowed` instead of `is less than
  limit allowed`
- Update tests to use go1.13 error checking instead of string matching
  since it is brittle and required all error messages to have a specific
  string inside of them which doesn't make sense in certain cases.

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25409